Allow customization of dependency comparison
Bundler uses `DepProxy` objects here which contain an underlying
depednency, and a platform. However, when comparing dependencies so that
spec groups with the same dependencies can be considered together,
bundler needs the platform ignored.

So, we'd like this comparision to be customizable, so that in bundler we
can do

```ruby
dependencies.map(&:dep) == other_depednencies.map(&:dep)
```

and get the optimization applied.